typ = "float" 인 attribute "double"로 변경
float type attribute를 query하면 부정확한 value를 리턴받음.

0.9가 나와야되는데 0.899999999342이런식...

+++ scripts/domino_edition/box/assembly_01/__init__.py
@@ -62,7 +62,7 @@ class Assembly01Data(piece.DData):
                         "multi": True},
             "origin_sub_ctl_count": {"typ": "long",
                                      "value": 2},
-            "origin_ctl_size": {"typ": "float",
+            "origin_ctl_size": {"typ": "double",
                                 "value": 1},
             "mode": {"typ": "enum",
                      "enumName": ["DEBUG", "WIP", "PUB"],
