+++ scripts/domino_edition/box/finger_01/__init__.py
@@ -46,7 +46,7 @@ class Finger01Data(piece.DData):
             "anchors": {"typ": "matrix",
                         "value": [self._m1, self._m2, self._m3, self._m4],
                         "multi": True},
-            "offset_pole_vec": {"typ": "float",
+            "offset_pole_vec": {"typ": "double",
                                 "value": 1,
                                 "channelBox": True},
             "offset_pole_vec_matrix": {"typ": "matrix",
@@ -209,19 +209,19 @@ class Finger01Rig(piece.Rig):
 
         self.auto_rot_attr = attribute.add(self.ik_ctl,
                                            "auto_rot",
-                                           typ="float",
+                                           typ="double",
                                            minValue=0,
                                            maxValue=1,
                                            defaultValue=1,
                                            keyable=True)
         self.rot_attr = attribute.add(self.ik_ctl,
                                       "rot",
-                                      typ="float",
+                                      typ="double",
                                       defaultValue=0,
                                       keyable=True)
         self.roll_attr = attribute.add(self.ik_ctl,
                                        "roll",
-                                       typ="float",
+                                       typ="double",
                                        defaultValue=0,
                                        keyable=True)
